move pages into src/, map @ to src root
