+++ src/pages/_app.tsx
@@ -5,15 +5,15 @@ import { config } from '@fortawesome/fontawesome-svg-core'
 import '@fortawesome/fontawesome-svg-core/styles.css'
 config.autoAddCss = false
 
-import '../public/_assets/fonts/karmilla.css'
-import '../public/_assets/fonts/katex/katex.css'
+import '../../public/_assets/fonts/karmilla.css'
+import '../../public/_assets/fonts/katex/katex.css'
 
 import { ThemeProvider, createGlobalStyle } from 'styled-components'
-import { theme } from '../src/theme'
+import { Theme, theme } from '@/theme'
 
 import * as Sentry from '@sentry/browser'
 
-import { version } from '../package.json'
+import { version } from '../../package.json'
 
 if (process.env.SENTRY_DSN !== undefined) {
   Sentry.init({
@@ -24,10 +24,12 @@ if (process.env.SENTRY_DSN !== undefined) {
 
 const FontFix = createGlobalStyle`
   h1,h2, main b {
-    letter-spacing: ${(props) => props.theme.defaults.boldLetterSpacing};
+    letter-spacing: ${(props) =>
+      (props.theme as Theme).defaults.boldLetterSpacing};
   }
   body {
-    letter-spacing: ${(props) => props.theme.defaults.regularLetterSpacing};
+    letter-spacing: ${(props) =>
+      (props.theme as Theme).defaults.regularLetterSpacing};
   }
 `
