+++ src/pages/api/frontend/[...slug].tsx
@@ -1,4 +1,4 @@
-import fetchContent from '../../../src/fetcher/serlo-api'
+import fetchContent from '@/fetcher/serlo-api'
 
 // Proxy the API Call as GET request to the frontend so that the ZEIT Now CDN is able to cache this
 // We use stale-while-revalidate for that, see also https://zeit.co/docs/v2/network/caching#stale-while-revalidate
