+++ src/theme.tsx
@@ -60,3 +60,5 @@ export const theme = {
     boldLetterSpacing: '-0.012em',
   },
 }
+
+export type Theme = typeof theme
